Upgrade to netcore 2.0. Build script for win-arm platform
Upgraded project and all NuGet packages to .netcore 2.0 to latest versions. This requires latest .NET Core SDK

Since the aim of this upgrade was to integrate project to Win10 IoT Core, running on rpi3, a script was added, which build targets ARM platform to support rpi3 hardware. To build on such platform, RestSharp.NetCore had to be changed to RestSharp to build without failures

Project, after these upgrades, is successfully up and running on latest Win10 IoT Core machine with rpi3